Split tl.copy lowering by backend
