+++ src/backend/cuda/CMakeLists.txt
@@ -109,6 +109,7 @@ file(GLOB TILE_LANG_CUDA_SRCS
   src/target/codegen_cutedsl.cc
   src/target/rt_mod_cuda.cc
   src/target/rt_mod_cutedsl.cc
+  src/backend/cuda/op/*.cc
 )
 list(APPEND TILE_LANG_SRCS ${TILE_LANG_CUDA_SRCS})
